requirements.txt
anthropic
pyaudio
python-dotenv
flask
speechrecognition
pydub

